+++ src/Infrastructure/Console/ElasticCreate.php
@@ -19,6 +19,8 @@ class ElasticCreate extends Command
 
     public function handle(ElasticIndexAdapter $adapter, IndexConfigurationRepositoryInterface $indexConfigurationRepository): int
     {
+        $this->warn('This command is deprecated since 2.4.0 and will be removed in 3.0. Use scout:index `name` instead.');
+
         $config = config('explorer');
         if (!$config) {
             $this->warn('There are no indices defined!');
@@ -28,6 +30,7 @@ class ElasticCreate extends Command
 
         foreach ($indexConfigurationRepository->getConfigurations() as $config) {
             $adapter->create($config);
+            $adapter->pointToAlias($config);
 
             $this->info('Created index ' . $config->getName());
         }