+++ tests/Unit/Operations/Bulk/BulkUpdateOperationTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace JeroenG\Explorer\Tests\Unit\Operations\Bulk;
 
 use JeroenG\Explorer\Application\Operations\Bulk\BulkUpdateOperation;
+use JeroenG\Explorer\Domain\IndexManagement\IndexConfiguration;
 use JeroenG\Explorer\Tests\Support\Models\TestModelWithoutSettings;
 use JeroenG\Explorer\Tests\Support\Models\TestModelWithPrepare;
 use JeroenG\Explorer\Tests\Support\Models\TestModelWithSettings;
@@ -14,13 +15,13 @@ class BulkUpdateOperationTest extends TestCase
 {
     public function test_it_builds_with_an_empty_command(): void
     {
-        $operation = new BulkUpdateOperation();
+        $operation = new BulkUpdateOperation(':searchable_as:');
         self::assertEquals([], $operation->build());
     }
 
     public function test_it_builds_with_a_model_command(): void
     {
-        $operation = new BulkUpdateOperation();
+        $operation = new BulkUpdateOperation(':searchable_as:');
         $operation->add(new TestModelWithoutSettings());
         self::assertEquals([
             ['index' => [ '_index' => ':searchable_as:', '_id' => ':scout_key:' ]],
@@ -30,7 +31,7 @@ class BulkUpdateOperationTest extends TestCase
 
     public function test_it_builds_with_multiple_model_command(): void
     {
-        $operation = new BulkUpdateOperation();
+        $operation = new BulkUpdateOperation(':searchable_as:');
 
         $operation->add(new TestModelWithoutSettings());
         $operation->add(new TestModelWithSettings());
@@ -49,7 +50,8 @@ class BulkUpdateOperationTest extends TestCase
      */
     public function test_it_builds_from_sources($input): void
     {
-        $operation = BulkUpdateOperation::from($input);
+        $indexConfiguration = IndexConfiguration::create(':searchable_as:', [], []);
+        $operation = BulkUpdateOperation::from($input, $indexConfiguration);
 
         self::assertEquals([
             ['index' => [ '_index' => ':searchable_as:', '_id' => ':scout_key:' ]],
@@ -66,7 +68,7 @@ class BulkUpdateOperationTest extends TestCase
 
     public function test_it_builds_with_preparation_of_model(): void
     {
-        $operation = new BulkUpdateOperation();
+        $operation = new BulkUpdateOperation(':searchable_as:');
         $operation->add(new TestModelWithPrepare());
         self::assertEquals([
             ['index' => [ '_index' => ':searchable_as:', '_id' => ':scout_key:' ]],