+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.4.0]
 
 ### Added
+- Index aliases, useful for zero downtime indexing
 - The scout:index and scout:delete commands are now implemented for Elasticsearch
 - A DocumentAdapterInterface with an adapter that only deals with documents
 - The Query String and Simple Query String syntax
@@ -16,9 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - The IndexAdapterInterface now only focuses on indices
+- The BulkOperation class now requires an instance of IndexConfigurationInterface
 
 ### Deprecated
 - If you rely on the old IndexAdapterInterface, use the DeprecatedElasticAdapterInterface instead of the IndexAdapterInterface to keep the functionality working as it previously was
+- The `elastic:create` and `elastic:delete` are deprecated in favour of `scout:index` and `scout:delete-index`
 
 ## [2.3.0]